Move GpuFirst to shim
