+++ shims/spark30/src/main/scala/com/nvidia/spark/rapids/shims/spark30/Spark30Shims.scala
@@ -22,7 +22,7 @@ import com.nvidia.spark.rapids._
 import org.apache.spark.sql.rapids.GpuTimeSub
 import org.apache.spark.sql.rapids.shims.spark30._
 
-//import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
+import org.apache.spark.sql.catalyst.expressions.aggregate.First
 import org.apache.spark.sql.execution.joins.{BuildLeft, BuildRight, BuildSide}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.execution._
@@ -32,8 +32,6 @@ import org.apache.spark.sql.execution.joins.ShuffledHashJoinExec
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.{CalendarInterval, UTF8String}
-
-
 
 class Spark30Shims extends SparkShims with Logging {
 
@@ -111,6 +109,17 @@ class Spark30Shims extends SparkShims with Logging {
         }
       }
     ),
+    GpuOverrides.expr[First](
+      "first aggregate operator",
+      (a, conf, p, r) => new ExprMeta[First](a, conf, p, r) {
+        val child: BaseExprMeta[_] = GpuOverrides.wrapExpr(a.child, conf, Some(this))
+        val ignoreNulls: BaseExprMeta[_] =
+          GpuOverrides.wrapExpr(a.ignoreNullsExpr, conf, Some(this))
+        override val childExprs: Seq[BaseExprMeta[_]] = Seq(child, ignoreNulls)
+
+        override def convertToGpu(): GpuExpression =
+          GpuFirst(child.convertToGpu(), ignoreNulls.convertToGpu())
+      }),
     )
   }
