+++ tests/pom.xml
@@ -32,12 +32,12 @@
 
     <properties>
         <slf4j.version>1.7.30</slf4j.version>
+        <spark.test.version>3.1.0-SNAPSHOT</spark.test.version>
     </properties>
     <profiles>
         <profile>
             <id>spark31</id>
             <properties>
-                <spark.version>3.1.0</spark.version>
             </properties>
         </profile>
     </profiles>
@@ -61,6 +61,7 @@
         <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-sql_${scala.binary.version}</artifactId>
+	    <version>${spark.test.version}</version>
         </dependency>
         <dependency>
             <groupId>org.scalatest</groupId>
